[Verilog] delay_buffer.v: fix reset logic (#389)
This commit fixes the reset logic of the first valid bit in the delay
buffer.

The reset logic before:

```verilog
  always @(posedge clk) begin
    if (ready_in | rst) begin
      regs[0] <= valid_in;
    end
  end
```

If we raise the rst pin when `valid_in = 1`, the delay buffer will reset
with one token (?!!).

This occurs super rarely and I discovered this when running model
checking on the circuit...

I fixed the reset logic like this:

```verilog
  always @(posedge clk) begin
    if (rst)
      regs[0] <= 0;
    else if (ready_in)
      regs[0] <= valid_in;
  end
```