[CfToHandshake] Add Argument Classification Based on Naming Convention in CfToHandshake (#384)
## Summary
This PR introduces logic to the cf-to-handshake pass that classifies
arguments of a called function based on their naming convention, in
preparation for proper support of multi-output MLIR operations. (see
#321 )

## Motivation
Currently, the conversion pass does not distinguish between function
arguments used as:
- Inputs
- Outputs
- Parameters

Proper classification is necessary to rewire outputs correctly and
support multi-output MLIR operations. Additionally, it allows for the
explicit specification of component parameters.

### Assumptions:
- All arguments of a func::FuncOp have a handshake.arg_name attribute
attached.
- Argument naming follows the expected convention
  - Names starting with input_* represent inputs.
  - Names starting with output_* represent outputs.
  - Names starting with parameter_* represent parameters.
- Only func::FuncOps (i.e., not yet lowered handshake functions) are
considered for argument classification and rewiring during the
cf-to-handshake conversion.
- The placeholder function is declared but not defined, so the compiler
treats it as an external function.


## Implementation
- Added argument classification logic inside
ConvertCalls::matchAndRewrite
- For each func::FuncOp refrenced by a call:
  1. Extracted the handshake.arg_name attribute of each argument
  2. Classified the argument into one of the following categories:
      - Inputs: If name starts with "input_*"
      - Outputs: If name starts with "output_*"
      - Parameters: If name starts with "parameter_*
- Names that fit in none of the above emit an llvm::errs() message and
fail the assertion.
3. Stored the Indices of the classified arguments inside one of three
vectors:
      - InstanceOpInputIndices
      - InstanceOpOutputIndices
      - InstanceOpParameterIndices
4. For each output argument index find all operations that consume its
value and store the mapping inside a dictionary called
OutputConnections.

This lays the groundwork for enabling dynamatic to handle multiple
output values via function arguments, differentiating parameters from
inputs/outputs and rewire output values to their consumers reliably
during cf-to-handshake conversion.

---------

Co-authored-by: Carmine50 <36476551+Carmine50@users.noreply.github.com>