[Buffer] Introduce explicit buffer type field (#317)
This PR introduces a more general buffer classification (see [Discussion
#269](https://github.com/EPFL-LAP/dynamatic/discussions/269)). For
detailed implementation principles, please refer to
[docs/Specs/Buffering.md](https://github.com/QinYuan2000/dynamatic/tree/buffertype/docs/Specs/Buffering.md).

You can view the results of the updated buffer pass on the
integration-test suite in
[Results.pdf](https://github.com/user-attachments/files/19508310/SP25_Buffertype_Data__cmp_internalbuf_.pdf).
For cycle count, the majority of benchmarks remain unchanged, with
differences in a few cases attributed to the inherent randomness of the
optimizer when handling large benchmarks. As for timing slack and area
utilization, the differences are partly due to optimizer randomness and
partly from a slight modification in the buffer placement strategy.
These changes generally lead to only small variations, and in many
cases, the metrics have improved.

To test this PR, simply run any integration test with the
--buffer-algorithm=fpga20 or --buffer-algorithm=fpl22 option enabled.

---

*Update (after the discussion with @murphe67)*: We will not remove
TimingAttr in this PR, as a preliminary step, we add BUFFER_TYPE
attribute, update the json with this attribute. We will later make a
project-wide change to make all buffer HDL instantiation use this
attribute. We might keep TimingAttr for other analysis pass (e.g., LSQ
sizing pass needs it for the buffer's latency).

---