[SMV] Add internal signals for Input Ports (#397)
We define internal signals for the input ports in SMV so that we can
output the ready signal of the input channels.

While in VHDL connecting the input channels to the internal units
already drives the ready signal, this is not true for SMV. In SMV we
need to explicitly define internal signals for the ready signals that
are sent outside the module. To do so we use the same approach as for
output channels with the `WriteModData::writeSignalAssignments`, but
inverting the directions (data and valid go in, ready goes out).

`WriteModData::writeSignalAssignments` is used in all HDLs, but the
internal signals for inputs are assigned only in SMV. If in the future
those signals will be needed also for VHDL and SV, they can be added
with very small changes.