[RTL] Suppress "Uncaught module" warning (#422)
Currently, during parameter analysis, we emit a warning like `Uncaught
module: ori` if a module isn't handled in the bitwidth or extra signal
extraction.
However, this warning isn't relevant for users relying on the existing
backend.
Moreover, it's already clear on the new backend whether a module is
analyzed, during Python generation. If the relevant parameter isn't
passed, Python emits an error.