[Buffers] Making Buffer Instantiation Consistent with the MILP Models (#429)
Previously, buffer placement was inconsistent with the MILP model.

In `FPGA20Buffers`, a ONE_SLOT_BREAK_R (TEHB) was used for buffers that
break no signal paths (i.e., transparent buffers) and have one slot.
This is incorrect. ONE_SLOT_BREAK_R breaks the READY signal and reduces
throughput compared to a truly transparent buffer. Therefore, we now
place FIFO_BREAK_NONE when the MILP assigns a one-slot buffer that
breaks no signals.

In `FPL22Buffers`, the buffer attributes (Break DV and Break R) are
currently inconsistent with the MILP constraints. As a result, any
buffer placement derived from them is not valid until the MILP is
corrected. This PR does not fix the MILP, but clarifies what the ideal
buffer placement should be.