[Buffers] Fix Buffering Constraints before Load/Store (#448)
In #356, we placed additional buffers at the input channels of
Load/Store units connected to the LSQ to resolve the latency mismatch
problem, as proposed in #388.

However, the inserted buffer does not need to be a transparent
FIFO—opaque buffers can also resolve the issue. This PR updates the
logic to allow any buffer type, preventing redundant insertion of a
transparent FIFO when an opaque buffer is already present.

---

Tests on a few benchmarks have proven its correctness and functionality.