[Experimental][SQ] Added Straight to Queue algorithm (#449)
This PR introduces in the branch `FTD-with-OOO` the code for straight to
the queue (FPGA'23).
The modifications to the codebase are minimal, except for the files
implementing the main logic of the algorithm:
- `dynamatic.cpp`: implements the logic to run pass some arguments to
`compile.sh`;
- `compile.sh`: implements the logic to integrate straight to the queue
in the flow. **Have a look at it before merging, make sure the flow is
what you expect to have**.

I run an integration test (`--fast-token-delivery` +
`--straight-to-queue`) using all the available tests from
`integration_test`. These are the failures:

- gemver/gemver_float (was failing in my thesis as well);
- lu (timeout, it used to take a lot of time);
- insertion_sort (was failing in my thesis as well);
- iterative_sqrt (this is currently failing with fast token delivery, so
I am not surprised it is here as well);
- test_stdint (this is currently failing with fast token delivery, so I
am not surprised it is here as well).