[SMV] Annotate Formal Properties (#402)
We add a new compiler pass called `AnnotateFormalProperties` that generates a database (json file) that contains all the information for the properties we want to check at the circuit level.

---

The properties in the database have the following information:
- ID: unique identifier of the property
- type: type of the property. For now we support only absence of
back-pressure (AOB) and valid-equivalence (VEQ)
- tag: specifies the purpose of the property. For now we support only
"optimization", meaning that these properties will be used to optimize
the circuit. In the future other tags can be included like "invariant"
or "error"
- check: contains the information returned by the model checker. It can
be either "unchecked" if the model checker hasn't been run yet, "true",
or "false".
- info: configurable parameter where all the information needed to write
the property in RTL can be stored