[ElasticMiter] Updated ElasticMiter docs and build scripts (#460)
I made several small updates to the `README.md` and the build scripts
(`./build.sh` and the CMake files).

The original `README.md` was unclear for new users, so I updated it.

Also, some headings were used incorrectly, and there was even a syntax
error that caused part of the content to be hidden on GitHub.

This raised a broader concern for me about the current review process in
Dynamatic. Was the file actually reviewed? I've noticed that some recent
PRs were merged very quickly, sometimes without any suggestions like
#449, even though they involved a significant amount of code. (CC:
@murphe67)

This can lead to issues for users later on. I actually ran into trouble
with the unreviewed `README.md` in the ElasticMiter project. It makes me
concerned about the quality of the C++ code in ElasticMiter now, since
C++ is generally harder to read than Markdown.

Based on the discussion in the Dynamatic meeting, I understand that PRs
aren't the right place to raise this concern, so I'll open a discussion
thread about it.

---

I also made some fixes to the build scripts.