[SMV] Property generation (#406)
We extend `export-rtl.cpp` to generate properties in SMV if a property
database is provided.

`export-rtl` doesn't make any decision on which properties to generate.
It just generates the properties based on the information provided by
the .json file that can be passed as an argument. The .json file can be
generated with the `AnnotateFormalProperties` pass in
https://github.com/EPFL-LAP/dynamatic/pull/402.