[SMV] Testbench generation (#459)
We generalized the testbench generation from the [elastic-miter
project](https://github.com/EPFL-LAP/dynamatic/issues/236) and we
extracted it into a support library so that it can be used also
elsewhere.

The testbench generation in SMV is needed every time we want to verify
something using a model checker. The testbench creates a main module
where the model checking starts and instantiates the model under test.
Additionally the testbench for elastic circuits needs to provide the
necessary input and output channels that respect the [handshaking
protocol](https://github.com/EPFL-LAP/dynamatic/issues/373). We provide
2 options for the outputs: we can either synchronize them with a join or
directly consume them with sinks.

---------

Co-authored-by: Jiahui Xu <48123265+Jiahui17@users.noreply.github.com>