Add missing gurobi-installed guards. (#490)
On my machine (up-to-date Arch install), dynamatic fails to build both
because of the GCC v15 issues (see #442) and because due to some
undefined references when building without gurobi installed.

From a cursory reading it seems that both `SubjectGraph.h` and
`SubjectGraph.cpp` depend on some data structures (such as `class Node`,
defined in `BlifReader.h`) which are only available when gurobi is
installed (for example, `BlifReader.h` contains if-def guards that
exclude it's content when `DYNAMATIC_GUROBI_NOT_INSTALLED` is defined).

Not knowing anything about this code - I assume that `SubjectGraph` is
only relevant when using dynamatic with gurobi?
If so, it is missing the required ifdef guards, which this PR 

If `SubjectGraph` is to be used also without gurobi, then the fix is not
as easy... and I don't know what it would be :)


Pinging @Carmine50 @cosmiclat05 since this was introduced in 491f81a /
#286 :):)

Cheers!