Update polygeist/LLVM to pick-up backported GCCv15 compat. fixes.
Fixes #442.

The version of LLVM dynamatic uses is the same version used by
Polygeist, which is a few years old.

As different linux distros have started shipping the new GCC v15, the
project no longer compiled correctly. This is because of some changes in
GCC's standard library headers, which used to include stdint.h (or
cstdint), but no longer do. Any files that relied on fixed-width
integers without themselves including the stdint may have compiled fine
with previous versions but fail with GCC v15.

https://gcc.gnu.org/gcc-15/porting_to.html

Also, new optimisations in GCC v15 have exposed some undefined behavior
in dynamatic that was not an issue before. These have been fixed in
separate PRs.

This PR changes polygeist commit to pickup a new different version of
LLVM, which is a fork of the correct version but with the relevant
patches back-ported.