[Timing Model] IR support for new combinational delays in pipelined units (#476)
# Summary 

## References 

This is a followup PR to
[PR#475](https://github.com/EPFL-LAP/dynamatic/pull/475), as part of the
wider framework outlined in
[issue#457](https://github.com/EPFL-LAP/dynamatic/issues/457). This PR
will assume you have already read PR#475 and understand the essential
data structures used, etc.

To summarize : PR#475 added support for multiple latencies for pipelined
units, and implemented logic to allow these to be accessed in all
dynamatic passes.


## Context

However, the current backend **can only generate a fixed version of a
given operator**, and the IR **does not carry any information which
could be used to identify which version is to be implemented**.

These two limitations will be the focus of this PR. 

# Assumptions

See PR#475

# Implementation

## IR alterations

### MLIR Dialect changes
In order to allow the internal delay to be selected in the buffer
placement pass, and then make it available to the backend, we add a new
attribute ```InternalDelay``` to the backend. Following the guidelines
layed out in the
[documentation](https://github.com/EPFL-LAP/dynamatic/blob/main/docs/Tutorials/AddNewComponent.md),
it is added as a standalone optional attr inside
[```ArithOps.td```](https://github.com/EPFL-LAP/dynamatic/blob/main/include/dynamatic/Dialect/Handshake/HandshakeArithOps.td)
rather than being added to the hw parameter blob.

### Buffer placement pass changes
During buffer placement pass, after the main body of the pass is
complete, we now iterate over the obtained IR. For every Op encountered,
we call a new function of the
[```DelayDepMetric```](https://github.com/EPFL-LAP/dynamatic/blob/main/include/dynamatic/Support/TimingModels.h#L96)
struc, which obtains, for the current targetCP, the appropriate delay,
and writes it into the IR.

### HW IR changes
Taking example on what is done for other attributes, minimal changes are
made to HandshaketoHW.cpp in order to appropriately declare the new
```InternalDelay``` parameter into the hw.mlir.

## Backend changes
The most substantial changes are made in the backend. 

The

---------

Co-authored-by: Carmine50 <36476551+Carmine50@users.noreply.github.com>