[TimingModel] Refactor internal delay to inteface attribute (#515)
This PR adds a handshake tablegen interface "InteralDelayInterface", which is now present on all arithmetic ops. The interface provides a dedicated getter and setter for a named StringAttr, with key "internal_delay". If the named StringAttr is not present, a StringAttr containing "0.0" is pulled from the MLIR context.

The intrinsic attribute has been removed, restoring the default builder to its previous state. The InternalDelayInterface is now used instead of the ArithOpInterface for adding the attribute. 

Finally, a dedicated if statement has been added to ModuleDiscriminator, since the interface applies to many different clauses of the original if statement.