[Timing] [Characterization] Add delay characterization scripts and documentation for dataflow units (#506)
The following PR adds a new script to run delay characterization of
dataflow units.

This is not a general script since many choices are statically present
in the code. However, it's a first attempt of generalizing delay
characterization.

The definitions of some units have been updated to be more
parser-friendly.

Additionally, the timing information has been updated using this new
characterization script.