[Floating Point Units] Add implementation attribute + new RTL and timing matching (#528)
Dynamatic now supports two sets of floating point units, with plans to
push for parameterized generation. Currently the two sets of floating
point units are separated by duplicating key config files.

Due to concerns about scalability, this PR moves to represent the
specific set of units that a floating point operation should draw from
in the IR itself.

In the future, this will hopefully allow us to expand the range of
floating point units that can be generated. As a first prototype, this
system allows a non-operation-specific set to be identified. Future
works will probably require operation-specific variants, which will
require changes.

This pass therefore adds:

1. A new enum MLIR attribute FPUImplEnumAttr which can be either
"flopoco" or "vivado".

2. An interface added to several operations which allows a pass to store
FPUImplEnumAttr in the operation using dedicated getters and setters.
The getter returns "vivado" if attribute has not been set on the
operation.

3. A new pass which takes a string option. The value of the option must
be symbolizable to a FPUImpl or the pass fails. The pass then iterates
over all operations with the new interface, and sets them to the enum
value specified in the option.

4. A new value is added to hw.params for operations which have the new
interface, which is a stringified version of the enum.

5. rtl-config can then match and pass based on that value, allowing us
to add multiple entries for each floating point operation, one entry per
set.

6. The timing model system is also updated. Previously, timing model
names were passed into the MLIR context as operation names, despite not
being part of the IR.

7. Now, timing model names are stored directly in a string map. Rather
than model names corresponding directly to operation names, the new
timing model names are based on both an operation's name and its
implementation, but only if they implement the new interface.