Add `mdbook` doc rendering. (#543)
As discussed, this sets-up `mdbook` rendering and github pages hosting.

Note that I have configured the CI to also push to github pages on
pull-request runs for testing, this will be changed to only publish on
actual merges/commits on the main branch.

The structure of the docs basically mimics what @A-Alosius prepared. I
suggest, for this PR, we only get the hosting infrastructure setup and
handle any changes to layout/structure, if there are any, separately.

This is not quite 100% done - but not marked as a `draft` to check if CI
works.

Co-authored-by: Eugen Bošnjak <eugen.bosnjak@epfl.ch>