[Sharing] Add resource sharing wrapper generator to beta backend (#553)
In order to remove the previous backend, we must support all units it
supported. The existing c++ sharing wrapper generator is incompatible
with the beta backend as it relies on the old dependancy system.

This PR adds a new python based sharing wrapper generator, as well as
the lazy fork to allow testing on the gemver kernel.