[FTD] Fix block identification from condition value (#564)
The block that has a particular condition value as its terminator
condition is not necessarily the same block that owns this value. We
need to make sure that the conditions of production and consumption
passed to the control dependence analysis that runs for fast token
delivery takes the block that has the condition as its terminator.

For instance, in reference to this #547, `%19` is produced in `^bb2` but
acts as the terminator condition of `^bb3`.

This PR fixes one part of fast token delivery that would return `^bb2`
for `%19` instead of returning `^bb3`, causing wrong results.