[FTD] Generalize Forward Control Dependencies Extraction (#566)
The `ControlDependenceAnalysis` pass implements the logic in
https://dl.acm.org/doi/pdf/10.1145/24039.24041 to identify the blocks
that every block is control dependent on, if any, and stores them in the
`allControlDeps` field. These dependencies include loop exits for blocks
contained in a loop.

FTD separates the logic of delivery across loops from that when the
producer and consumer are at the same loop level to avoid token count
mismatches. It does so by extracting a subset of the dependencies in
`allControlDeps` to `forwardControlDeps` and uses the latter in studying
the conditions of production and consumption when the producer and
consumer happen to be at the same loop level.

Currently, the condition of extracting `forwardControlDeps` from
`allControlDeps` is too constrained and fails with the following CFG,
for instance, if we have a producer in `^bb1` to a consumer in `^bb3`.
`^bb3` should be control dependent both on `^bb1` and `^bb2`, but with
the current implementation `^bb2` is skipped because it is a loop tail.

<img width="312" height="513" alt="iterative_sqrt_CFG"
src="https://github.com/user-attachments/assets/ad82f732-4fa4-46cb-a34e-a17ae033db13"
/>

The purpose of this PR is to generalize this extraction by simply
counting all dependencies that come before the block in the CFG as a
`forwardControlDeps`.

P.S. Need to test it on all integration-tests then will mark the PR as
non draft.