[CI][Integration] Add example verilog tests (#573)
This adds a way to run benchmarks with the Verilog backend in the
GoogleTest test suite by setting a boolean argument in
`runIntegrationTest`. There is also an example test case that runs
benchmarks with Verilog, but it is commented out because a significant
number of benchmarks are still not working properly with that backend.