[CI] Separate various CI steps (#563)
This PR changes the Actions setup so that the workflow is not as
monolithic as it is right now. It is still a work in progress, but in
the end, the formatting check will be one separate job, ~~build and
check-dynamatic tests will be another,~~ (this would not be worth it
because build needs to run on the same runner, separating the jobs would
require maintaining state and artifacts between jobs) and integration
tests will be another. The goal is to make the setup clearer and easier
to understand, while also providing better feedback for the person
creating the PR.

This also resolves any issues with leftover state from previous runs by
emptying the workspace contents before all other steps.