[FTD] Generalizing the identification of a while loop +  the logic for accounting for a mux condition (#568)
This PR does three things:

1. generalizes the logic of identifying a while loop,
2. generalizes the logic of deciding when the mux condition should be
considered in the construction of the expressions for conditions of
production and consumption. Specifically, we only count muxes that
correspond to a Gamma gate,
3. removes an extra redundant `rewriter.setInsertionPoint` in
`addGsaGates` that can be problematic if we decide to allocate the gate
in a block different from the block of its result.

The changes are tested to not break any of the working integration
tests.