Remove set-polygeist-path (#575)
The set-polygeist-path frontend command is an awkward solution to an
external polygeist build (which I currently use) as it requires the
command to be added to many scripts through dynamatic, but is a per-user
setting.

I have removed this command and restored the previous symbolic links,
which work better. I have also altered the include file setup, and so
have added an extra symbolic link to the polygeist headers, and then
altered the call to polygeist in compile.sh