Support Loops in dfsAllPaths When Start and End Are the Same Block (#559)
This pull request introduces changes to the `dfsAllPaths` function to
support loops, specifically the case where the start and end blocks are
the same.

In the SSA-to-GSA conversion, when we have a phi node with more than two
inputs, we want to break it into a tree of gates. For loops, this often
means identifying:
- A unique incoming value from outside the loop (initial value)
- A unique value from inside the loop

However, if either of these values is not unique, we build a tree of
gates to combine them into a single gate output.
(Note: for this to work properly, we also made tweaks to the GSA
Analysis pass, but we will open a separate PR for them.)

Let’s consider the following
[case](https://github.com/EPFL-LAP/dynamatic/pull/369#:~:text=An%20example%20loop%20with%20a%20merge%20having%203%2Dor%2Dmore%20inputs%3A):
```
module {
  func.func @eg2(%arg1: memref<200xi32> {handshake.arg_name = "b"}) {
    %c1_i32 = arith.constant {handshake.name = "constant1"} 1 : i32
    cf.br ^bb1(%c1_i32 : i32) {handshake.name = "br0"}
  ^bb1(%0: i32):  // 3 preds: ^bb0, ^bb1, ^bb2
    %c1_i32_0 = arith.constant {handshake.name = "constant2"} 1 : i32
    %c198_i32 = arith.constant {handshake.name = "constant3"} 198 : i32
    %1 = arith.addi %0, %c1_i32_0 {handshake.name = "addi0"} : i32
    %2 = arith.index_cast %1 {handshake.name = "index_cast0"} : i32 to index
    memref.store %c1_i32_0, %arg1[%2] {handshake.mem_interface = #handshake.mem_interface<MC>, handshake.name = "store0"} : memref<200xi32>
    %3 = arith.addi %0, %c1_i32_0 {handshake.name = "addi1"} : i32
    %4 = arith.cmpi slt, %0, %c198_i32 {handshake.name = "cmpi1"} : i32
    cf.cond_br %4, ^bb1(%3 : i32), ^bb2(%1, %0 : i32, i32) {handshake.name = "cond_br0"}
  ^bb2(%5: i32, %6: i32):  // pred: ^bb1
    %c198_i32_1 = arith.constant {handshake.name = "constant4"} 198 : i32
    %7 = arith.cmpi ult, %6, %c198_i32_1 {handshake.name = "cmpi2"} : i32
    cf.cond_br %7, ^bb1(%5 : i32), ^bb3 {handshake.name = "cond_br1"}
  ^bb3:  // pred: ^bb2
    return {handshake.name = "return0"}
  }
}
```
<img width="173" height="388" alt="image"
src="https://github.com/user-attachments/assets/b6c50a1a-99e3-44ee-b0a9-df3c35facd34"
/>

As shown in the control flow graph, `bb1` has 3 predecessors: `bb0`,
`bb1`, and `bb2`.
The value from `bb0` is the initial one (outside the loop), while values
from `bb1` and `bb2` are computed in the loop.
We want to build:
- A mux to choose between the loop value and the outside value
- Another mux to distinguish between values from `bb1` and `bb2` (the
loop-internal values)

For the second mux, we need to compute the conditions under which each
incoming value is selected. This requires exploring paths like `{bb1,
bb2, bb1}` and `{bb1, bb1}`. But the existing `dfsAllPaths` would not
allow revisiting `bb1` twice and exit early when `start == end`, missing
these valid loop paths.

What This PR Does:

1. Supports revisiting end block, to allow loop paths like {bb1, bb1}
and {bb1, bb2, bb1}.
2. Limits "end" revisits using "endVisitCount" to avoid infinite
traversal (max 2 times).
3. Ensures that the path has progressed at least one edge (path.size() >
1), so we allow proper condition extraction later in the analysis.

---------

Co-authored-by: Eugen Bošnjak <eugen.bosnjak@epfl.ch>
Co-authored-by: ayaelakhras <ayaelakhras@aucegypt.edu>