[CfToHandshake] Refactoring "convertMemoryOps" (#551)
> [!NOTE]
> This is a necessary change to support internal arrays in Dynamatic
(tracked in
https://github.com/EPFL-LAP/dynamatic/tree/feature/add-memory-dep-analysis)

The `CfToHandshake` conversion pass currently uses
`convertNonEntryRegionTypes` to rewrite the signatures of each BB from
built-in types (e.g., i32) to handshake types (e.g., channel<i32>).

However, this approach does not update the operands of the users of the
block arguments, which turns all the OpOperands of the ops that consume
the BB arguments into `<<UNKNOWN SSA VALUES>>`.

Some other methods use getMemRef() to retrieve the ops, but since
everything becomes `<<UNKNOWN SSA VALUES>>`, the current implementation
had to use some hacky way to recover the actual getMemRef().

This PR simplifies the convoluted logic by directly mutating the BB
signature instead of creating a set of new signatures.

Although this is NOT recommended in the official repo (mutating the
Value types might lead to invalid IR),

https://github.com/llvm/llvm-project/blob/6d90715019d54376523163a7e1d588e1068cfca2/mlir/include/mlir/IR/Value.h#L110-L115
I argue that this method is safe, as it is used in a conversion pass
where every operation will be rebuilt in the end.