[FTD][GSA] Support for Multi-Input Loops, Path Distinction, and Missing-Phi Duplicate-checking (#567)
This PR addresses three limitations in the current GSAAnalysis
implementation.

**1. Multi-input loop**

Currently, GSA does not support loops where a phi has 3 or more inputs.

To fix this issue, we separate the operands of such phi nodes into two
groups:
- Initial values (from outside the loop)
- Loop values (from inside the loop)

If either group has more than one operand, we insert an extra phi gate
to resolve it. The outputs of these intermediate phis are then connected
as inputs to a mu gate. These phi gates will later be converted into
gamma gates as usual.

All related changes are in `convertPhiToMu`.



**2. Distinguishing multiple values from the same producer block**

If a block produces more than one value that feeds the same phi argument
via different paths, the current algorithm cannot distinguish them. This
is because `findAllPaths` only considers paths from the common dominator
to the phi block, without tracking which value flows along which path.

Example (iterative_sqrt with a small change):

```
int iterative_sqrt(in_int_t n) {
 int low = 0, high = n, mid;
 while (low <= high) {
   mid = ((low + high) >> 1);
   if (mid * mid == n) {
     high = mid;
     break;
   }
   else if (mid * mid < n) {
     low = mid + 1;
   }
   else {
     high = mid - 1;
   }
 }
 return high;
}
```

```
module {
  func.func @iterative_sqrt(%arg0: i32 {handshake.arg_name = "n"}) -> i32 {
    %c0_i32 = arith.constant {handshake.name = "constant0"} 0 : i32
    %true = arith.constant {handshake.name = "constant3"} true
    cf.br ^bb1(%arg0, %c0_i32, %true : i32, i32, i1) {handshake.name = "br0"}
  ^bb1(%0: i32, %1: i32, %2: i1):  // 4 preds: ^bb0, ^bb2, ^bb4, ^bb5
    %3 = arith.cmpi sle, %1, %0 {handshake.name = "cmpi0"} : i32
    %4 = arith.andi %3, %2 {handshake.name = "andi0"} : i1
    cf.cond_br %4, ^bb2(%0, %1 : i32, i32), ^bb6 {handshake.name = "cond_br0"}
  ^bb2(%5: i32, %6: i32):  // pred: ^bb1
    %c1_i32 = arith.constant {handshake.name = "constant4"} 1 : i32
    %7 = arith.addi %6, %5 {handshake.name = "addi0"} : i32
    %8 = arith.shrsi %7, %c1_i32 {handshake.name = "shrsi0"} : i32
    %9 = arith.muli %8, %8 {handshake.name = "muli0"} : i32
    %10 = arith.cmpi ne, %9, %arg0 {handshake.name = "cmpi1"} : i32
    %11 = arith.cmpi sle, %6, %5 {handshake.name = "cmpi2"} : i32
    %12 = arith.cmpi sgt, %6, %5 {handshake.name = "cmpi3"} : i32
    %13 = arith.andi %11, %10 {handshake.name = "andi1"} : i1
    %14 = arith.cmpi eq, %9, %arg0 {handshake.name = "cmpi4"} : i32
    %15 = arith.ori %13, %12 {handshake.name = "ori0"} : i1
    cf.cond_br %14, ^bb1(%8, %6, %15 : i32, i32, i1), ^bb3 {handshake.name = "cond_br1"}
  ^bb3:  // pred: ^bb2
    %16 = arith.cmpi slt, %9, %arg0 {handshake.name = "cmpi5"} : i32
    cf.cond_br %16, ^bb4, ^bb5 {handshake.name = "cond_br2"}
  ^bb4:  // pred: ^bb3
    %c1_i32_0 = arith.constant {handshake.name = "constant5"} 1 : i32
    %17 = arith.addi %8, %c1_i32_0 {handshake.name = "addi1"} : i32
    cf.br ^bb1(%5, %17, %15 : i32, i32, i1) {handshake.name = "br1"}
  ^bb5:  // pred: ^bb3
    %c-1_i32 = arith.constant {handshake.name = "constant6"} -1 : i32
    %18 = arith.addi %8, %c-1_i32 {handshake.name = "addi2"} : i32
    cf.br ^bb1(%18, %6, %15 : i32, i32, i1) {handshake.name = "br7"}
  ^bb6:  // pred: ^bb1
    return {handshake.name = "return0"} %0 : i32
  }
}

```
<img width="300" height="504" alt="image"
src="https://github.com/user-attachments/assets/4572b60b-c8be-4afd-b7b2-a28aa6f85bd0"
/>

- Block bb2 produces both %5 and %8.
- %8 is sent to arg0 bb1 via bb2.
- %5 is sent to arg0 bb1 via bb4.

Assuming bb2 is the common dominator of the phi for arg0 bb1, with the
current code, both paths {bb2, bb1} and {bb2, bb3, bb4} are considered
for both %5 and %8, which is incorrect.

After the fixes, each gate input records its sender predecessors as a
set of blocks.
This set is filled when a new block argument is detected (value or
missing phi). It is also updated in `isAlreadyPresent`, both for values
and missing-phi cases.
Later, in `convertPhiToGamma`, when all paths are found, we filter them
to only keep the paths that match the correct sender set.


**3. Improved duplicate-checking for missing-phi operands**

Previously, when creating gate inputs, missing-phi operands were not
checked for duplication. This caused duplicated entries with identical
conditions that couldn't be resolved in `expandGammaTree`.

The `isAlreadyPresent` function is improved to also check missing-phi
operands. Two missing-phi operands are considered equal if:
- They target the same phi argument, and
- They originate from the same producer block.

(The actual value is not compared, which may need refinement in the
future.)

---------

Co-authored-by: Aya ElAkhras <ayatallah.elakhras@epfl.ch>
Co-authored-by: Eugen Bošnjak <eugen.bosnjak@epfl.ch>