[Backend] Refactor arithmetic units in VHDL beta backend (#552)
In order to remove our need for 4 different backends, we should expand
the capabilities of the beta backend to encompass all of them.

This PR changes how we generate all of our arithmetic units in the VHDL
beta backend, such that it now contains all features of the current VHDL
backend, including the work on floating point units by @Carmine50 and
@KillianMcCourt.

It also converts our floating point subtraction unit to be a direct
wrapper around our adder.