[Backend] Fixed CMerge-with-data Impl (#625)
## Problem

The previous implementation of control merge with data (currently only
used with extra signals) contained a bug: CMerge includes an internal
TEHB to manage the handshaking protocol, but the data path was directly
connected without buffering, leading to a token mismatch.

## Solution

I didn't fix this issue in the old Verilog and VHDL backends, as they
are deprecated and will soon be replaced by the beta backend. Instead I
raised an error.

In the beta backend, RTL is no longer separated into dataless and
data-carrying variants. This simplifies the implementation, by using the
`data` utility function.